requirements.txt
langchain
langchain-core
langchain-google-genai
python-dotenv
deep-translator